consolidate helper funcs in _common/azure_utils.py
