+++ python/ray/llm/_internal/common/utils/cloud_filesystem/pyarrow_filesystem.py
@@ -149,28 +149,9 @@ class PyArrowFileSystem(BaseCloudFileSystem):
                 credential=DefaultAzureCredential(),
             )
         except Exception as e:
-            _is_credential_error = False
-            try:
-                from azure.core.exceptions import ClientAuthenticationError
+            from ray._common.azure_utils import handle_azure_credential_error
 
-                _is_credential_error = isinstance(e, ClientAuthenticationError)
-            except ImportError:
-                pass
-            if not _is_credential_error:
-                try:
-                    from azure.identity import CredentialUnavailableError
-
-                    _is_credential_error = isinstance(e, CredentialUnavailableError)
-                except ImportError:
-                    pass
-            if _is_credential_error:
-                raise RuntimeError(
-                    f"Azure credential error accessing storage account "
-                    f"{azure_storage_account_name}.\n"
-                    "Credentials may be expired or misconfigured. "
-                    "Try: `az login`\n"
-                    "Or set: AZURE_CLIENT_ID, AZURE_CLIENT_SECRET, AZURE_TENANT_ID"
-                ) from e
+            handle_azure_credential_error(e, resource_type="Azure storage")
             raise
 
         # Wrap with PyArrow's PyFileSystem for compatibility
